Separate knn settings defs from util methods
Initial commit of refactoring change around KNNSettings. The goal of the
refactor is to tie setting definitions with related resources and make
utility class lighter weight.

In future commits, I will clean up the KNNSettingsUtil interface and
move settings to the classes that best define them.
