+++ src/main/java/org/opensearch/knn/index/KNNSettingsDefinitions.java
@@ -1,0 +1,240 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package org.opensearch.knn.index;
+
+import org.opensearch.OpenSearchParseException;
+import org.opensearch.common.settings.Setting;
+import org.opensearch.core.common.unit.ByteSizeValue;
+import org.opensearch.common.unit.TimeValue;
+import org.opensearch.monitor.jvm.JvmInfo;
+
+import java.util.HashMap;
+import java.util.Map;
+
+import static org.opensearch.common.settings.Setting.Property.Dynamic;
+import static org.opensearch.common.settings.Setting.Property.IndexScope;
+import static org.opensearch.common.settings.Setting.Property.NodeScope;
+import static org.opensearch.common.unit.MemorySizeValue.parseBytesSizeValueOrHeapRatio;
+import static org.opensearch.knn.index.KNNSettings.parseknnMemoryCircuitBreakerValue;
+import static org.opensearch.knn.index.KNNSettings.percentageAsFraction;
+import static org.opensearch.knn.index.KNNSettings.percentageAsString;
+
+/**
+ * This class simply defines all of the settings and their names for k-NN
+ */
+public class KNNSettingsDefinitions {
+    private static final int INDEX_THREAD_QTY_MAX = 32;
+
+    /**
+     * Settings name
+     */
+    public static final String KNN_SPACE_TYPE = "index.knn.space_type";
+    public static final String KNN_ALGO_PARAM_M = "index.knn.algo_param.m";
+    public static final String KNN_ALGO_PARAM_EF_CONSTRUCTION = "index.knn.algo_param.ef_construction";
+    public static final String KNN_ALGO_PARAM_EF_SEARCH = "index.knn.algo_param.ef_search";
+    public static final String KNN_ALGO_PARAM_INDEX_THREAD_QTY = "knn.algo_param.index_thread_qty";
+    public static final String KNN_MEMORY_CIRCUIT_BREAKER_ENABLED = "knn.memory.circuit_breaker.enabled";
+    public static final String KNN_MEMORY_CIRCUIT_BREAKER_LIMIT = "knn.memory.circuit_breaker.limit";
+    public static final String KNN_CIRCUIT_BREAKER_TRIGGERED = "knn.circuit_breaker.triggered";
+    public static final String KNN_CACHE_ITEM_EXPIRY_ENABLED = "knn.cache.item.expiry.enabled";
+    public static final String KNN_CACHE_ITEM_EXPIRY_TIME_MINUTES = "knn.cache.item.expiry.minutes";
+    public static final String KNN_PLUGIN_ENABLED = "knn.plugin.enabled";
+    public static final String KNN_CIRCUIT_BREAKER_UNSET_PERCENTAGE = "knn.circuit_breaker.unset.percentage";
+    public static final String KNN_INDEX = "index.knn";
+    public static final String MODEL_INDEX_NUMBER_OF_SHARDS = "knn.model.index.number_of_shards";
+    public static final String MODEL_INDEX_NUMBER_OF_REPLICAS = "knn.model.index.number_of_replicas";
+    public static final String MODEL_CACHE_SIZE_LIMIT = "knn.model.cache.size.limit";
+    public static final String ADVANCED_FILTERED_EXACT_SEARCH_THRESHOLD = "index.knn.advanced.filtered_exact_search_threshold";
+
+    /**
+     * Default setting values
+     */
+    public static final String INDEX_KNN_DEFAULT_SPACE_TYPE = "l2";
+    public static final Integer INDEX_KNN_DEFAULT_ALGO_PARAM_M = 16;
+    public static final Integer INDEX_KNN_DEFAULT_ALGO_PARAM_EF_SEARCH = 100;
+    public static final Integer INDEX_KNN_DEFAULT_ALGO_PARAM_EF_CONSTRUCTION = 100;
+    public static final Integer KNN_DEFAULT_ALGO_PARAM_INDEX_THREAD_QTY = 1;
+    public static final Integer KNN_DEFAULT_CIRCUIT_BREAKER_UNSET_PERCENTAGE = 75;
+    public static final Integer KNN_DEFAULT_MODEL_CACHE_SIZE_LIMIT_PERCENTAGE = 10; // By default, set aside 10% of the JVM for the limit
+    public static final Integer KNN_MAX_MODEL_CACHE_SIZE_LIMIT_PERCENTAGE = 25; // Model cache limit cannot exceed 25% of the JVM heap
+    public static final String KNN_DEFAULT_MEMORY_CIRCUIT_BREAKER_LIMIT = "50%";
+
+    public static final Integer ADVANCED_FILTERED_EXACT_SEARCH_THRESHOLD_DEFAULT_VALUE = -1;
+
+    /**
+     * Settings Definition
+     */
+
+    public static final Setting<String> INDEX_KNN_SPACE_TYPE = Setting.simpleString(
+        KNN_SPACE_TYPE,
+        INDEX_KNN_DEFAULT_SPACE_TYPE,
+        new KNNSettings.SpaceTypeValidator(),
+        IndexScope,
+        Setting.Property.Deprecated
+    );
+
+    /**
+     * M - the number of bi-directional links created for every new element during construction.
+     * Reasonable range for M is 2-100. Higher M work better on datasets with high intrinsic
+     * dimensionality and/or high recall, while low M work better for datasets with low intrinsic dimensionality and/or low recalls.
+     * The parameter also determines the algorithm's memory consumption, which is roughly M * 8-10 bytes per stored element.
+     */
+    public static final Setting<Integer> INDEX_KNN_ALGO_PARAM_M_SETTING = Setting.intSetting(
+        KNN_ALGO_PARAM_M,
+        INDEX_KNN_DEFAULT_ALGO_PARAM_M,
+        2,
+        IndexScope,
+        Setting.Property.Deprecated
+    );
+
+    /**
+     *  ef or efSearch - the size of the dynamic list for the nearest neighbors (used during the search).
+     *  Higher ef leads to more accurate but slower search. ef cannot be set lower than the number of queried nearest neighbors k.
+     *  The value ef can be anything between k and the size of the dataset.
+     */
+    public static final Setting<Integer> INDEX_KNN_ALGO_PARAM_EF_SEARCH_SETTING = Setting.intSetting(
+        KNN_ALGO_PARAM_EF_SEARCH,
+        INDEX_KNN_DEFAULT_ALGO_PARAM_EF_SEARCH,
+        2,
+        IndexScope,
+        Dynamic
+    );
+
+    /**
+     * ef_constrution - the parameter has the same meaning as ef, but controls the index_time/index_accuracy.
+     * Bigger ef_construction leads to longer construction(more indexing time), but better index quality.
+     */
+    public static final Setting<Integer> INDEX_KNN_ALGO_PARAM_EF_CONSTRUCTION_SETTING = Setting.intSetting(
+        KNN_ALGO_PARAM_EF_CONSTRUCTION,
+        INDEX_KNN_DEFAULT_ALGO_PARAM_EF_CONSTRUCTION,
+        2,
+        IndexScope,
+        Setting.Property.Deprecated
+    );
+
+    public static final Setting<Integer> MODEL_INDEX_NUMBER_OF_SHARDS_SETTING = Setting.intSetting(
+        MODEL_INDEX_NUMBER_OF_SHARDS,
+        1,
+        1,
+        Setting.Property.NodeScope,
+        Setting.Property.Dynamic
+    );
+
+    public static final Setting<Integer> MODEL_INDEX_NUMBER_OF_REPLICAS_SETTING = Setting.intSetting(
+        MODEL_INDEX_NUMBER_OF_REPLICAS,
+        1,
+        0,
+        Setting.Property.NodeScope,
+        Setting.Property.Dynamic
+    );
+
+    public static final Setting<Integer> ADVANCED_FILTERED_EXACT_SEARCH_THRESHOLD_SETTING = Setting.intSetting(
+        ADVANCED_FILTERED_EXACT_SEARCH_THRESHOLD,
+        ADVANCED_FILTERED_EXACT_SEARCH_THRESHOLD_DEFAULT_VALUE,
+        IndexScope,
+        Setting.Property.Dynamic
+    );
+
+    public static final Setting<ByteSizeValue> MODEL_CACHE_SIZE_LIMIT_SETTING = new Setting<>(
+        MODEL_CACHE_SIZE_LIMIT,
+        percentageAsString(KNN_DEFAULT_MODEL_CACHE_SIZE_LIMIT_PERCENTAGE),
+        (s) -> {
+            ByteSizeValue userDefinedLimit = parseBytesSizeValueOrHeapRatio(s, MODEL_CACHE_SIZE_LIMIT);
+
+            // parseBytesSizeValueOrHeapRatio will make sure that the value entered falls between 0 and 100% of the
+            // JVM heap. However, we want the maximum percentage of the heap to be much smaller. So, we add
+            // some additional validation here before returning
+            ByteSizeValue jvmHeapSize = JvmInfo.jvmInfo().getMem().getHeapMax();
+            if ((userDefinedLimit.getKbFrac() / jvmHeapSize.getKbFrac()) > percentageAsFraction(
+                KNN_MAX_MODEL_CACHE_SIZE_LIMIT_PERCENTAGE
+            )) {
+                throw new OpenSearchParseException(
+                    "{} ({} KB) cannot exceed {}% of the heap ({} KB).",
+                    MODEL_CACHE_SIZE_LIMIT,
+                    userDefinedLimit.getKb(),
+                    KNN_MAX_MODEL_CACHE_SIZE_LIMIT_PERCENTAGE,
+                    jvmHeapSize.getKb()
+                );
+            }
+
+            return userDefinedLimit;
+        },
+        Setting.Property.NodeScope,
+        Setting.Property.Dynamic
+    );
+
+    /**
+     * This setting identifies KNN index.
+     */
+    public static final Setting<Boolean> IS_KNN_INDEX_SETTING = Setting.boolSetting(KNN_INDEX, false, IndexScope);
+
+    /**
+     * index_thread_quantity - the parameter specifies how many threads the nms library should use to create the graph.
+     * By default, the nms library sets this value to NUM_CORES. However, because ES can spawn NUM_CORES threads for
+     * indexing, and each indexing thread calls the NMS library to build the graph, which can also spawn NUM_CORES threads,
+     * this could lead to NUM_CORES^2 threads running and could lead to 100% CPU utilization. This setting allows users to
+     * configure number of threads for graph construction.
+     */
+    public static final Setting<Integer> KNN_ALGO_PARAM_INDEX_THREAD_QTY_SETTING = Setting.intSetting(
+        KNN_ALGO_PARAM_INDEX_THREAD_QTY,
+        KNN_DEFAULT_ALGO_PARAM_INDEX_THREAD_QTY,
+        1,
+        INDEX_THREAD_QTY_MAX,
+        NodeScope,
+        Dynamic
+    );
+
+    public static final Setting<Boolean> KNN_CIRCUIT_BREAKER_TRIGGERED_SETTING = Setting.boolSetting(
+        KNN_CIRCUIT_BREAKER_TRIGGERED,
+        false,
+        NodeScope,
+        Dynamic
+    );
+
+    public static final Setting<Double> KNN_CIRCUIT_BREAKER_UNSET_PERCENTAGE_SETTING = Setting.doubleSetting(
+        KNN_CIRCUIT_BREAKER_UNSET_PERCENTAGE,
+        KNN_DEFAULT_CIRCUIT_BREAKER_UNSET_PERCENTAGE,
+        0,
+        100,
+        NodeScope,
+        Dynamic
+    );
+    /**
+     * Dynamic settings
+     */
+    public static Map<String, Setting<?>> dynamicCacheSettings = new HashMap<String, Setting<?>>() {
+        {
+            /**
+             * KNN plugin enable/disable setting
+             */
+            put(KNN_PLUGIN_ENABLED, Setting.boolSetting(KNN_PLUGIN_ENABLED, true, NodeScope, Dynamic));
+
+            /**
+             * Weight circuit breaker settings
+             */
+            put(KNN_MEMORY_CIRCUIT_BREAKER_ENABLED, Setting.boolSetting(KNN_MEMORY_CIRCUIT_BREAKER_ENABLED, true, NodeScope, Dynamic));
+            put(
+                KNN_MEMORY_CIRCUIT_BREAKER_LIMIT,
+                new Setting<>(
+                    KNNSettingsDefinitions.KNN_MEMORY_CIRCUIT_BREAKER_LIMIT,
+                    KNNSettingsDefinitions.KNN_DEFAULT_MEMORY_CIRCUIT_BREAKER_LIMIT,
+                    (s) -> parseknnMemoryCircuitBreakerValue(s, KNNSettingsDefinitions.KNN_MEMORY_CIRCUIT_BREAKER_LIMIT),
+                    NodeScope,
+                    Dynamic
+                )
+            );
+
+            /**
+             * Cache expiry time settings
+             */
+            put(KNN_CACHE_ITEM_EXPIRY_ENABLED, Setting.boolSetting(KNN_CACHE_ITEM_EXPIRY_ENABLED, false, NodeScope, Dynamic));
+            put(
+                KNN_CACHE_ITEM_EXPIRY_TIME_MINUTES,
+                Setting.positiveTimeSetting(KNN_CACHE_ITEM_EXPIRY_TIME_MINUTES, TimeValue.timeValueHours(3), NodeScope, Dynamic)
+            );
+        }
+    };
+}
